backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Full-Stack Multi-User Todo Web Application - Backend"
authors = [{name = "Developer", email = "developer@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlmodel>=0.0.20",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "emails>=0.6",
    "jinja2>=3.1.0",
    "alembic>=1.13.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
    "psycopg2-binary>=2.9.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true